Refactor response handling
